implement portfolio screen & improve permission requester

+++ app/src/main/java/com/seftian/bnicasestudies/ui/screens/payment_detail/PaymentDetailScreen.kt
@@ -1,12 +1,9 @@
 package com.seftian.bnicasestudies.ui.screens.payment_detail
 
-import android.util.Log
-import android.widget.Toast
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -32,7 +29,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -103,7 +99,7 @@ fun PaymentDetailScreen(
                         id = it.transactionId,
                         sourceBank = it.bank,
                         merchantName = it.merchantName,
-                        trxAmount = it.transactionAmount
+                        trxAmount = it.transactionAmount.toLong()
                     )
                     viewModel.insertTrxToLocal(trx)
                 }
